app/requirements.txt
aiohttp==3.10.5
arq==0.26.0
asyncpg==0.29.0
asyncpraw==7.7.1
click==8.1.7
cryptography==43.0.0
fastapi==0.112.2
hiredis==3.0.0
openai==1.42.0
PyJWT==2.9.0
python-decouple==3.8
python-multipart==0.0.9
SQLAlchemy==2.0.32
sqlalchemy-utils==0.41.2
textblob==0.18.0.post0
validators==0.33.0
